fix: Adjust white space/indentation
This commit fixes the many inconsistancies white whitespace and indentation in the main SASS file. The file now consistently uses two spaces as a form of indentation for each sass property of a particular element/selector. Line breaks are also provided for each individual css property, instead of sometimes having all properties for a particular element on one line and sometimes not.