fix: remove row padding from portfolio grid
This commit removes the excessive padding/height Chrome adds to each of the items in the grid of portfolio projects on the front page of the website. The grid itself previously had a defined grid-auto-rows property of 50%, which was ignored in Firefox but not Chrome. This property has been removed, because there is already an explict row-gap define (as well as column gap) of 1.5em, and no need to define additional padding or set heights on each row.