refactor: reorganize elements into BEM-friendly structure
This commit introduces many changes the naming and organization structure of the CSS. Instead of styling root elements, which have a very high level of specificity, we will add classes to HTML5 elements according to BEM naming rules, as possible.

Some content on the website is included with root elements (such as paragraphs) and cannot be given BEM classes without adding additional markup to markdown files.

In the future, this may be a necessary change to make so that the css for the website does not mix and match BEM and root element styles. I need to do additional research about how this is handled by other Jekyll projects.