fix: add image class for smaller images
This commit adds a class for use with article images that are not wide enough to be resized to 130% of the width of the article's text column. It's a hack that requires the use of !important to override the high-level of specificity of the img selector.

I really dislike hacks: however, in this instance I'm adding a small amount of code to fix an individual edge case. Doing so enables all other instances of images in markdown files to not have to be marked up with additional css.